ARTEMIS-2320 Fix IntLongMath errorprone warning
Expression of type int may overflow before being assigned to a long
    (see https://errorprone.info/bugpattern/IntLongMath)

+++ artemis-core-client/src/main/java/org/apache/activemq/artemis/core/client/impl/ClientMessageImpl.java
@@ -387,7 +387,7 @@ public class ClientMessageImpl extends CoreMessage implements ClientMessageInter
          if (isLargeMessage()) {
             return getBodyBuffer().writerIndex();
          } else {
-            return getBodyBuffer().writerIndex() - BODY_OFFSET;
+            return (long) getBodyBuffer().writerIndex() - BODY_OFFSET;
          }
       }
